pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "branch-monkey"
version = "0.1.0"
description = "Git for humans - visual, safe, and beginner-friendly"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Branch Monkey Contributors" }
]
keywords = ["git", "tui", "version-control", "beginner-friendly"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gitpython>=3.1.40",
    "textual>=0.47.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
monkey = "branch_monkey.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/branch-monkey"
Documentation = "https://github.com/yourusername/branch-monkey#readme"
Repository = "https://github.com/yourusername/branch-monkey"

[tool.hatch.build.targets.wheel]
packages = ["branch_monkey"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
